+++ src/admin/wata/dto/find-wata.dto.ts
@@ -5,7 +5,7 @@ import { WataLabelType, WataRequiredValuesType } from '../interface/wata.type';
 import {
   QueryValidEnumArray,
   QueryValidNumberArray,
-} from '../../../common/decorators/dto.decorator';
+} from '../../../common/public.decorator.ts/dto.decorator';
 
 // docs - query number[] issue : https://dev.to/avantar/validating-numeric-query-parameters-in-nestjs-gk9
