+++ src/publish-wata/publish-wata.controller.ts
@@ -9,7 +9,7 @@ import {
 import { PublishWataService } from './publish-wata.service';
 import { ApiBearerAuth, ApiOperation, ApiTags } from '@nestjs/swagger';
 import { CacheKey, CacheTTL } from '@nestjs/cache-manager';
-import { Public } from 'src/common/decorators/public.decorator';
+import { Public } from 'src/common/public.decorator.ts/public.decorator';
 import {
   CACHE_TTL,
   HttpCacheInterceptor,
